+++ packages/bluesky/lib/src/services/extensions/at_uri.dart
